Rename `resources` to `contexts`
