+++ hotham/src/schedule_functions/physics_step.rs
@@ -1,4 +1,4 @@
-use crate::resources::PhysicsContext;
+use crate::contexts::PhysicsContext;
 
 // TODO: We may want to adjust this so that the tick time is synced with OpenXR
 pub fn physics_step(physics_context: &mut PhysicsContext) {
